Move and update file structure
