+++ src/routes/(app)/namespaces/+layout.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import { updated } from '$app/stores';
-  import Header from './_header.svelte';
+  import Header from '../../_header.svelte';
   import Notifications from '$lib/components/notifications.svelte';
   import Banners from '$lib/components/banner/banners.svelte';
   import { ErrorBoundary } from '$lib/components/error-boundary';
