+++ src/routes/(app)/namespaces/+page.svelte
