+++ src/routes/(app)/namespaces/[namespace]/+page.svelte
