+++ src/routes/(app)/namespaces/[namespace]/+page.ts
