+++ src/routes/(app)/namespaces/[namespace]/archival/+page.svelte
