+++ src/routes/(app)/namespaces/[namespace]/schedules/+layout.svelte
