+++ src/routes/(app)/namespaces/[namespace]/schedules/+page.svelte
