+++ src/routes/(app)/namespaces/[namespace]/schedules/[schedule]/+page.svelte
