+++ src/routes/(app)/namespaces/[namespace]/schedules/new/+page.svelte
