+++ src/routes/(app)/namespaces/[namespace]/workflows/[workflow]/[run]/history/compact/+page.svelte
