+++ src/routes/(app)/namespaces/[namespace]/workflows/[workflow]/[run]/stack-trace/+page.svelte
