+++ src/routes/(import)/import/events/[namespace]/[workflow]/[run]/history/compact/+page.svelte
