+++ src/routes/(login)/login/+layout.svelte
@@ -1,0 +1,5 @@
+<script lang="ts">
+  import '../../../app.css';
+</script>
+
+<slot />
